feat: add JavaScript search functionality
Added a function that handles the processing of the search text
whenever a key is released while focused on the search input element.

Added a function that searches thru the fruits list to see if any fruit
contains a given string of characters and returns an array of fruits.

Also changed the fruit list name to fruits, to better suit what it is.