feat: show suggestions
Added a showSuggestions function that will replace all of the possible
search results in the suggestions "ul" element, whether or not there is
any, with a new set of results.

Created a helper function separateStringForBolding to split a string
into an array to more easily put a piece of the string into a bold
element.

Fixed or added a feature in searchHandler that will remove the list of
possible results when there is no text in the search bar.