feat: add css
Added meta info on HTML page to have it adapt to mobile devices.

In the CSS, changed the selectors related to "ul" to
".search-container ul" to not affect possible future "ul"s.

In the CSS, the html, body, and classes of search-container had to have
a height of 100% to center the search bar and suggestions on the
webpage.