fix: hide suggestions when off-clicking
After typing into the search bar and bringing up the results, when
clicking anywhere besides the search bar or suggestions, the
suggestions drop-down does not disappear.  This is not the typical
expected behavior.

To fix this, an event listener was added to the document body that
checks if the event target is not the search bar or suggestions and
hides the suggestions if it isn't.

A hideSuggestions function was created to support this.

The searchHandler function was modified to first handle if there is a
class called "hide".  This was done for performance, since it is
unnecessary to recreate the same suggestions list that was hidden.

Also added an event listener on the search bar input to watch for
clicks and run searchHandler.  This is done to conform to expected
behavior to have the suggestions drop down upon clicking when there is
text in the search bar.

A new class was added to each suggestion / search result list item to
make it easier to specifically select the results, as opposed to
traversing parents and children of elements.  As a result, the CSS was
also refactored to use the class name.