test: add tests
Tests were only added for the functions that did not interact with the
DOM for the sake of time.

The HTML test webpage that displays the test results had to copy over
the actual webpage, because the actual JavaScript immediately uses a
HTML element.  Otherwise the Jasmine framework will throw an error,
which would be due to the actual JavaScript trying to access a
nonexistent HTML element.  I currently do not know of a better way to
do this without polluting production HTML code with test code.