fix: adjust the showing and hiding of suggestions
In the future, there can be instances where the suggestions are hidden
but a user would be typing in the search bar.  After the first key up,
the old suggestions list will be displayed instead of a new list for
the now different input text.

The checking of the class "hide" in the searchHandler function is now
removed.  Upon releasing keys, it will now always show suggestions.

The displaying and hiding of suggestions now only occurs when clicking
away from and on the search bar and container.